refactor(server): :art: Split `index.ts` into `app.ts` and `index.ts`
This is for easier testing
